🔥 chore: rename services/github to services/GithubService
