+++ src/controllers/deployController/deployDomain.ts
@@ -10,7 +10,7 @@ import { createDeploymentDebug } from "../../utils/createDebug";
 
 import { RRType } from "@aws-sdk/client-route-53";
 import { RecordSetResponse } from "../../types/custom";
-import { deleteRepoWebhook } from "../../services/github/client";
+import { deleteRepoWebhook } from "../../services/GithubService/client";
 
 const deployDomain = catchAsync(async (req, res, next) => {
   const debug = createDeploymentDebug(Config.CLIENT_OPTIONS.debug);
