+++ src/controllers/updateController/updateDeployment.ts
@@ -2,7 +2,7 @@ import mongoose from "mongoose";
 
 import Config from "../../config";
 
-import { getHeadCommitMessage } from "../../services/github/client";
+import { getHeadCommitMessage } from "../../services/GithubService/client";
 import runUpdateDeploymentCommands from "../../services/deploy/cli/runUpdateDeploymentCommands";
 
 import catchAsync from "../../utils/asyncHandler";
